Updated using header file

+++ XBee.h
@@ -5,5 +5,5 @@
 int receiveIntXbee();
 int sendIntXbee(int value);
 int sendShortXbee(short value);
-int receiveShortXbee();
+short receiveShortXbee();
 int initializeXBEE();
